feat: add event command
